Add files via upload
For most of the examples, I found a source that not only writes about the utterance and why it is problematic but the actual source of the person saying/posting it, but for some it was not possible. I added dates of utterance where I could find them but for some (Wahlprogramm for example), there was no date.