Examples of German dogwhistles used by politicians
In this excel sheet, examples of possible dog whistles used by German politicians were collected. Most of the examples also have a quote in which the term was used. Some of the quotes might not be considered dog whistles while the word itself, used in a slightly different example could possibly be an example (testing necessary).