Add files via upload
Those are the two questionnaires for Experiment 2. Each contains 4 dogwhistles (dw) and 4 minimal pairs (mp) and 4 controls (Aussage 3 and Aussage 9 are not plausibly deniable, and Aussage 6 and 11 are plausibly deniable.)
In "Plausible deniability (1_2)" the dw are Aussagen 1, 5, 7, 10 . The mp are Aussagen 2, 4, 8, 12.
In "Plausible deniability (2_2)" the dw are Aussagen 2, 4, 8, 12. The mp are Aussagen 1, 5, 7, 10.